feat(db/repos): extract solarEstimates, upgradeHistory, newsCache, backupHistory from misc.ts
PR1 of the misc.ts split (#3962 Phase 3). Move 4 global pure-Drizzle domains
into their own repository files:
- solarEstimates.ts — 3 methods (upsertSolarEstimate, getRecent, getInRange)
- upgradeHistory.ts — 11 methods + IN_PROGRESS_STATUSES private field
- newsCache.ts — 4 methods (getNewsCache, saveNewsCache, getUserNewsStatus, saveUserNewsStatus)
- backupHistory.ts — 7 methods (insertBackupHistory, getBackupHistoryList, getBackupByFilename, deleteBackupHistory, countBackups, getOldestBackups, getBackupStats)

Each domain: verbatim method move, per-domain test file with SQLite/PG/MySQL
harness, domain-only DDL in test setup. misc.ts retains remaining 9 domains;
unused drizzle imports (gte/lte/lt) removed.
